+++ src/main/java/bitcopark/library/repository/jwt/RefreshTokenBlackListRepository.java
@@ -1,8 +1,7 @@
-package bitcopark.library.jwt;
+package bitcopark.library.repository.jwt;
 
 import bitcopark.library.entity.jwt.RefreshTokenBlackList;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.transaction.annotation.Transactional;
 
 public interface RefreshTokenBlackListRepository extends JpaRepository<RefreshTokenBlackList, Long> {
